+++ todo-app/src/index.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import TodoApp from "./components/TodoApp"
+
+ReactDOM.render(
+  <TodoApp />,
+  document.getElementById('root')
+);
